iterative assembler: extend algorithm
Before we looked at all sequences containing the current start/end of the contig and took the base most likely to follow that value.  Now we weight those sequences by how well they agree with the current contig before the bit we matched on.

This was much too slow in python, so just the scoring bit I moved to C